LayoutTests/imported/w3c/web-platform-tests/pointerevents/pointerevent_boundary_events_before_pointerrawupdate.https-expected.txt


FAIL Pointer boundary events should be fired before `pointerrawupdate` assert_equals: expected "pointerover, pointerenter, pointerrawupdate, pointermove" but got "pointerover, mouseover, pointerenter, mouseenter, pointermove"
